back/requirements.txt
amqp==5.1.1
appdirs==1.4.4
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
Django==4.1.5
django-cors-headers==3.13.0
django-filter==22.1
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
executing==1.2.0
fonttools==4.38.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
ipython==8.8.0
jedi==0.18.2
Jinja2==3.1.2
kiwisolver==1.4.4
kombu==5.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mne==0.21.0
numpy==1.24.1
packaging==23.0
pandas==1.1.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
plotly==5.12.0
pooch==1.6.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pyparsing==3.0.9
pyTelegramBotAPI==4.9.0
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.4.2
requests==2.28.2
scipy==1.10.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tenacity==8.1.0
tqdm==4.64.1
traitlets==5.8.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
zipp==3.11.0
